+++ app/src/main/kotlin/com/yapt/planttracker/ui/screens/addplant/AddEditPlantScreen.kt
@@ -318,6 +318,7 @@ fun AddEditPlantScreen(
             DormancyWindowSetting(
                 startMonth = viewModel.dormancyStartMonth,
                 endMonth = viewModel.dormancyEndMonth,
+                dormantWateringIntervalDays = viewModel.dormantWateringIntervalDays,
                 onWindowChange = viewModel::setDormancyWindow
             )